+++ mypy/build.py
@@ -121,6 +121,7 @@ def build(sources: List[BuildSource],
           bin_dir: str = None,
           pyversion: Tuple[int, int] = defaults.PYTHON3_VERSION,
           custom_typing_module: str = None,
+          implicit_any: bool = False,
           report_dirs: Dict[str, str] = {},
           flags: List[str] = None,
           python_path: bool = False) -> BuildResult:
@@ -140,6 +141,7 @@ def build(sources: List[BuildSource],
         directories; if omitted, use '.' as the data directory
       pyversion: Python version (major, minor)
       custom_typing_module: if not None, use this module id as an alias for typing
+      implicit_any: if True, add implicit Any signatures to all functions
       flags: list of build options (e.g. COMPILE_ONLY)
     """
     flags = flags or []
@@ -186,6 +188,7 @@ def build(sources: List[BuildSource],
                            pyversion=pyversion, flags=flags,
                            ignore_prefix=os.getcwd(),
                            custom_typing_module=custom_typing_module,
+                           implicit_any=implicit_any,
                            reports=reports)
 
     # Construct information that describes the initial files. __main__ is the
@@ -330,6 +333,7 @@ class BuildManager:
                  flags: List[str],
                  ignore_prefix: str,
                  custom_typing_module: str,
+                 implicit_any: bool,
                  reports: Reports) -> None:
         self.data_dir = data_dir
         self.errors = Errors()
@@ -339,6 +343,7 @@ class BuildManager:
         self.pyversion = pyversion
         self.flags = flags
         self.custom_typing_module = custom_typing_module
+        self.implicit_any = implicit_any
         self.reports = reports
         self.semantic_analyzer = SemanticAnalyzer(lib_path, self.errors,
                                                   pyversion=pyversion)
@@ -804,7 +809,8 @@ class UnprocessedFile(State):
         num_errs = self.errors().num_messages()
         tree = parse.parse(source_text, fnam, self.errors(),
                            pyversion=self.manager.pyversion,
-                           custom_typing_module=self.manager.custom_typing_module)
+                           custom_typing_module=self.manager.custom_typing_module,
+                           implicit_any=self.manager.implicit_any)
         tree._fullname = self.id
         if self.errors().num_messages() != num_errs:
             self.errors().raise_error()
