+++ mypy/nodes.py
@@ -12,8 +12,6 @@ from mypy.lex import Token
 import mypy.strconv
 from mypy.visitor import NodeVisitor
 from mypy.util import dump_tagged, short_type
-from mypy.options import Options
-
 
 
 class Context(metaclass=ABCMeta):
@@ -336,16 +334,7 @@ class FuncItem(FuncBase):
         arg_kinds = [arg.kind for arg in self.arguments]
         self.max_pos = arg_kinds.count(ARG_POS) + arg_kinds.count(ARG_OPT)
         self.body = body
-        if typ is not None:
-            self.type = typ
-        elif Options.get_options().implicit_any:
-            from mypy.types import (UnboundType, CallableType)
-            self.type = CallableType([UnboundType('Any', []) for _ in arguments],
-                                     arg_kinds,
-                                     [a.variable.name() for a in arguments],
-                                     UnboundType('Any', []),
-                                     None)
-
+        self.type = typ
         self.expanded = []
 
         self.min_args = 0
